requirements.txt
pyserial==3.4
python-aqi==0.6.1
